+++ api/src/main/java/de/hoelscherp/invoice/api/api/InvoicesApiImpl.java
@@ -15,32 +15,32 @@ public class InvoicesApiImpl implements InvoicesApi {
     private final InvoiceService service;
 
     @Override
-    public ResponseEntity<Invoice> createInvoice(Integer customerId, Invoice invoice) {
-        Invoice saved = service.create(customerId.longValue(), invoice);
+    public ResponseEntity<Invoice> createInvoice(Long customerId, Invoice invoice) {
+        Invoice saved = service.create(customerId, invoice);
         return ResponseEntity.ok(saved);
     }
 
     @Override
-    public ResponseEntity<Void> deleteInvoice(Integer id) {
-        service.delete(id.longValue());
+    public ResponseEntity<Void> deleteInvoice(Long id) {
+        service.delete(id);
         return ResponseEntity.noContent().build();
     }
 
     @Override
-    public ResponseEntity<Invoice> getInvoice(Integer id) {
-        Invoice record = service.get(id.longValue());
+    public ResponseEntity<Invoice> getInvoice(Long id) {
+        Invoice record = service.get(id);
         return ResponseEntity.ok(record);
     }
 
     @Override
-    public ResponseEntity<List<Invoice>> getInvoicesByCustomer(Integer customerId) {
-        List<Invoice> invoices = service.getByCustomer(customerId.longValue());
+    public ResponseEntity<List<Invoice>> getInvoicesByCustomer(Long customerId) {
+        List<Invoice> invoices = service.getByCustomer(customerId);
         return ResponseEntity.ok(invoices);
     }
 
     @Override
-    public ResponseEntity<Invoice> updateInvoice(Integer id, Invoice invoice) {
-        Invoice updatedRecord = service.update(id.longValue(), invoice);
+    public ResponseEntity<Invoice> updateInvoice(Long id, Invoice invoice) {
+        Invoice updatedRecord = service.update(id, invoice);
         return ResponseEntity.ok(updatedRecord);
     }
 }
